static/requirements.txt
seleniumbase
pandas
pyyaml
pytest
openpyxl
